just a few suggestions
Sam, I have not been able to efficiently find a way to work with the
csv_download object, could we have this actually download this object to
a local directory then have a subsequent line read from this location?

I could add this to a loop so we have all csv, overwritten one the
district changes..

Let me know.